pyproject.toml
[project]
name = "pdf-deinjection"
version = "0.1.0"
description = "Windows desktop app that rasterizes PDFs into image-only PDFs to remove hidden prompt injection content."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "customtkinter>=5.2.0",
    "pillow>=10.0.0",
    "pymupdf>=1.24.0",
    "tkinterdnd2>=0.4.3",
    "pyinstaller>=6.0.0",
]

[tool.uv]
package = false